refactor: move NewPvcName func to k8sutil package
