+++ pkg/migrate/migrate.go
@@ -11,6 +11,7 @@ import (
 	"text/tabwriter"
 	"time"
 
+	"github.com/replicatedhq/pvmigrate/pkg/k8sutil"
 	corev1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
@@ -173,7 +174,7 @@ func copyAllPVCs(ctx context.Context, w *log.Logger, clientset k8sclient.Interfa
 	w.Printf("\nCopying data from %s PVCs to %s PVCs\n", sourceSCName, destSCName)
 	for ns, nsPvcs := range matchingPVCs {
 		for _, nsPvc := range nsPvcs {
-			sourcePvcName, destPvcName := nsPvc.Name, newPvcName(nsPvc.Name)
+			sourcePvcName, destPvcName := nsPvc.Name, k8sutil.NewPvcName(nsPvc.Name)
 			w.Printf("Copying data from %s (%s) to %s in %s\n", sourcePvcName, nsPvc.Spec.VolumeName, destPvcName, ns)
 
 			err := copyOnePVC(ctx, w, clientset, ns, sourcePvcName, destPvcName, rsyncImage, verboseCopy, waitTime, rsyncFlags)
@@ -471,7 +472,7 @@ func getPVCs(ctx context.Context, w *log.Logger, clientset k8sclient.Interface, 
 	w.Printf("\nCreating new PVCs to migrate data to using the %s StorageClass\n", destSCName)
 	for ns, nsPvcs := range matchingPVCs {
 		for _, nsPvc := range nsPvcs {
-			newName := newPvcName(nsPvc.Name)
+			newName := k8sutil.NewPvcName(nsPvc.Name)
 
 			desiredPV, ok := pvsByName[nsPvc.Spec.VolumeName]
 			if !ok {
@@ -574,24 +575,6 @@ func validateStorageClasses(ctx context.Context, w *log.Logger, clientset k8scli
 	}
 	w.Printf("\nMigrating data from %s to %s\n", sourceSCName, destSCName)
 	return nil
-}
-
-const nameSuffix = "-pvcmigrate"
-
-// if the length after adding the suffix is more than 63 characters, we need to reduce that to fit within k8s limits
-// pruning from the end runs the risk of dropping the '0'/'1'/etc of a statefulset's PVC name
-// pruning from the front runs the risk of making a-replica-... and b-replica-... collide
-// so this removes characters from the middle of the string
-// TODO: refactor to k8sutil package
-func newPvcName(originalName string) string {
-	candidate := originalName + nameSuffix
-	if len(candidate) <= 253 {
-		return candidate
-	}
-
-	// remove characters from the middle of the string to reduce the total length to 63 characters
-	newCandidate := candidate[0:100] + candidate[len(candidate)-153:]
-	return newCandidate
 }
 
 // get a PV, apply the selected mutator to the PV, update the PV, use the supplied validator to wait for the update to show up
@@ -975,9 +958,9 @@ func swapPVs(ctx context.Context, w *log.Logger, clientset k8sclient.Interface, 
 	if err != nil {
 		return fmt.Errorf("failed to get original PVC %s in %s: %w", pvcName, ns, err)
 	}
-	migratedPVC, err := clientset.CoreV1().PersistentVolumeClaims(ns).Get(ctx, newPvcName(pvcName), metav1.GetOptions{})
+	migratedPVC, err := clientset.CoreV1().PersistentVolumeClaims(ns).Get(ctx, k8sutil.NewPvcName(pvcName), metav1.GetOptions{})
 	if err != nil {
-		return fmt.Errorf("failed to get migrated PVC %s in %s: %w", newPvcName(pvcName), ns, err)
+		return fmt.Errorf("failed to get migrated PVC %s in %s: %w", k8sutil.NewPvcName(pvcName), ns, err)
 	}
 
 	// mark PVs used by both originalPVC and migratedPVC as to-be-retained
@@ -1018,10 +1001,10 @@ func swapPVs(ctx context.Context, w *log.Logger, clientset k8sclient.Interface, 
 	if err != nil {
 		return fmt.Errorf("failed to delete original PVC %s in %s: %w", pvcName, ns, err)
 	}
-	w.Printf("Deleting migrated-to PVC %s in %s to release the PV\n", newPvcName(pvcName), ns)
-	err = clientset.CoreV1().PersistentVolumeClaims(ns).Delete(ctx, newPvcName(pvcName), metav1.DeleteOptions{})
+	w.Printf("Deleting migrated-to PVC %s in %s to release the PV\n", k8sutil.NewPvcName(pvcName), ns)
+	err = clientset.CoreV1().PersistentVolumeClaims(ns).Delete(ctx, k8sutil.NewPvcName(pvcName), metav1.DeleteOptions{})
 	if err != nil {
-		return fmt.Errorf("failed to delete migrated-to PVC %s in %s: %w", newPvcName(pvcName), ns, err)
+		return fmt.Errorf("failed to delete migrated-to PVC %s in %s: %w", k8sutil.NewPvcName(pvcName), ns, err)
 	}
 
 	// wait for the deleted PVCs to actually no longer exist
@@ -1030,10 +1013,10 @@ func swapPVs(ctx context.Context, w *log.Logger, clientset k8sclient.Interface, 
 	if err != nil {
 		return fmt.Errorf("failed to ensure deletion of original PVC %s in %s: %w", pvcName, ns, err)
 	}
-	w.Printf("Waiting for migrated-to PVC %s in %s to finish deleting\n", newPvcName(pvcName), ns)
-	err = waitForDeletion(ctx, clientset, newPvcName(pvcName), ns)
+	w.Printf("Waiting for migrated-to PVC %s in %s to finish deleting\n", k8sutil.NewPvcName(pvcName), ns)
+	err = waitForDeletion(ctx, clientset, k8sutil.NewPvcName(pvcName), ns)
 	if err != nil {
-		return fmt.Errorf("failed to ensure deletion of migrated-to PVC %s in %s: %w", newPvcName(pvcName), ns, err)
+		return fmt.Errorf("failed to ensure deletion of migrated-to PVC %s in %s: %w", k8sutil.NewPvcName(pvcName), ns, err)
 	}
 
 	// remove claimrefs from original and migrated-to PVs
